requirements.txt
streamlit
Pillow
numpy
pydicom
